Rendering + Playability
Game is playable and it, technically speaking, renders.
Still need to add GUI elements, make it so the game actually ends, scales with screen size, and if I have time, can be restarted without refreshing the page.
Feeling pretty good about this, but i'm gonna have to put a lot of work in